@@ -4,8 +4,8 @@
   "private": true,
   "homepage": "https://zkindest.github.io/react-game-of-life",
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -27,10 +27,10 @@
   },
   "devDependencies": {
     "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@vitejs/plugin-react": "^1.1.4",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
+    "@vitejs/plugin-react": "^4.0.0",
     "typescript": "^4.5.4",
-    "vite": "^2.7.12"
+    "vite": "^4.2.0"
   }
 }
