@@ -1,14 +1,16 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import './index.css'
-import App from './App'
 import reportWebVitals from './reportWebVitals'
+import { createRoot } from 'react-dom/client'
+import App from './App'
 
-ReactDOM.render(
+const container = document.getElementById('root') as HTMLElement
+
+const root = createRoot(container) // createRoot(container!) if you use TypeScript
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
 
 // If you want to start measuring performance in your app, pass a function
